new_PPI_WEB/requirements.txt
streamlit
biopython
transformers
torch
numpy
matplotlib
scikit-learn
